Move RtcRtpSendParameters into separate module
